split helpers into separate files, remove unused randomHex helper, mark package as side effect free
