+++ package.json
@@ -34,5 +34,6 @@
     "test": "jest --detect-open-handles",
     "coverage": "jest --coverage",
     "prepublishOnly": "yarn && make"
-  }
+  },
+  "sideEffects": false
 }
