+++ src/helpers/findlnurl.ts
@@ -1,0 +1,9 @@
+/** @format */
+
+export function findlnurl(bodyOfText: string): string | null {
+  let res = /,*?((lnurl)([0-9]{1,}[a-z0-9]+){1})/.exec(bodyOfText.toLowerCase())
+  if (res) {
+    return res[1]
+  }
+  return null
+}
